"0.3brokencaching", 2006-06-25 22:43:55
It's probably broken. ¯\_(ツ)_/¯

I'm going to include the generated files in this because a) they're in
the ZIP archives b) Git practice is generally less principled about this
than Subversion/CVS were c) I suspect the number of people willing to
deal with random Perl scripts is dwindling these days.