"0.3allvehicles", 2006-06-26 00:32:23
As it says, now including data for all the normal vehicles.

I think one of the things that meant I never considered this mutator
"1.0" and released it back in the day (like...fourteen years ago) was
that this is a lousy static list, rather than being able to detect which
vehicles and vehicle weapons exist, so it'd work with other mutators
that added any. :C